// src/app/api-cinekeep/model/pageResponseWatchlistItemResponse.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WatchlistItemResponse } from './watchlistItemResponse';


export interface PageResponseWatchlistItemResponse { 
    content?: Array<WatchlistItemResponse>;
    page?: number;
    size?: number;
    totalElements?: number;
    totalPages?: number;
}